requirements.txt
# ==========================
# Core Framework
# ==========================
# Django framework
django==5.1.4

# ==========================
# REST APIs
# ==========================
# Django REST framework for building APIs
djangorestframework==3.15.2
# Simple JWT authentication for Django REST framework
djangorestframework_simplejwt==5.3.1
# This package provides JSON renderers and parsers that automatically
# convert between camelCase and snake_case. It ensures that API responses
# follow camelCase (preferred in JavaScript) while Django processes data
# internally using snake_case (Python standard). Ideal for frontend-backend
# integration in REST APIs.
djangorestframework-camel-case==1.4.2
# DJ-REST-AUTH for user authentication and registration
dj-rest-auth==7.0.0
# DRF-YASG for generating Swagger UI for APIs
drf-yasg

# ==========================
# Social Authentication
# ==========================
# DJ-REST-AUTH with social authentication support
dj-rest-auth[with_social]

# ==========================
# Environment Variables Management
# ==========================
# Python dotenv for loading environment variables from .env files
python-dotenv

# ==========================
# Database
# ==========================
# Parser for database URLs
dj_database_url==2.3.0
# MySQL client for Python
mysqlclient==2.1.1

# ==========================
# Task Queues and Schedulers
# ==========================
# Celery for asynchronous task queues
celery==5.3.1
# Django Celery Beat for periodic tasks scheduling
django-celery-beat==2.7.0
# Real-time monitoring tool for Celery tasks
flower==2.0.1
# Redis client for Python
redis

# ==========================
# Production WSGI Server
# ==========================
# Gunicorn for running the application in production
gunicorn==20.1.0